@@ -8,8 +8,8 @@
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>2.7.17</version>
-		<relativePath /> <!-- lookup parent from repository -->
+		<version>3.2.2</version>
+		<relativePath />
 	</parent>
 
 	<groupId>jkml</groupId>
@@ -17,8 +17,8 @@
 	<version>0.0.1-SNAPSHOT</version>
 
 	<properties>
-		<java.version>11</java.version>
-		<json-path.version>2.8.0</json-path.version>
+		<java.version>17</java.version>
+		<json-path.version>2.9.0</json-path.version>
 		<org.json.version>20231013</org.json.version>
 		<commons-cli.version>1.6.0</commons-cli.version>
 		<jacoco.version>0.8.11</jacoco.version>
@@ -41,11 +41,6 @@
 			<groupId>org.slf4j</groupId>
 			<artifactId>slf4j-nop</artifactId>
 			<scope>runtime</scope>
-		</dependency>
-
-		<dependency>
-			<groupId>net.minidev</groupId>
-			<artifactId>json-smart</artifactId>
 		</dependency>
 
 		<dependency>
@@ -75,31 +70,7 @@
 
 	<build>
 		<plugins>
-			<plugin>
-				<groupId>pl.project13.maven</groupId>
-				<artifactId>git-commit-id-plugin</artifactId>
-				<configuration>
-					<generateGitPropertiesFile>true</generateGitPropertiesFile>
-					<failOnNoGitDirectory>false</failOnNoGitDirectory>
-					<offline>true</offline>
-					<excludeProperties>
-						<excludeProperty>git.build.host</excludeProperty>
-						<excludeProperty>git.build.user.*</excludeProperty>
-						<excludeProperty>git.commit.message.*</excludeProperty>
-						<excludeProperty>git.commit.user.*</excludeProperty>
-						<excludeProperty>git.local.*</excludeProperty>
-						<excludeProperty>git.remote.*</excludeProperty>
-					</excludeProperties>
-				</configuration>
-				<executions>
-					<execution>
-						<id>default-revision</id>
-						<goals>
-							<goal>revision</goal>
-						</goals>
-					</execution>
-				</executions>
-			</plugin>
+
 			<plugin>
 				<groupId>org.jacoco</groupId>
 				<artifactId>jacoco-maven-plugin</artifactId>
@@ -119,6 +90,23 @@
 					</execution>
 				</executions>
 			</plugin>
+
+			<plugin>
+				<groupId>io.github.git-commit-id</groupId>
+				<artifactId>git-commit-id-maven-plugin</artifactId>
+				<configuration>
+					<generateGitPropertiesFile>true</generateGitPropertiesFile>
+					<failOnNoGitDirectory>false</failOnNoGitDirectory>
+					<offline>true</offline>
+					<includeOnlyProperties>
+						<includeOnlyProperty>git.build.time</includeOnlyProperty>
+						<includeOnlyProperty>git.commit.id</includeOnlyProperty>
+						<includeOnlyProperty>git.commit.time</includeOnlyProperty>
+						<includeOnlyProperty>git.dirty</includeOnlyProperty>
+					</includeOnlyProperties>
+				</configuration>
+			</plugin>
+
 			<plugin>
 				<groupId>org.springframework.boot</groupId>
 				<artifactId>spring-boot-maven-plugin</artifactId>
@@ -128,8 +116,8 @@
 					</layers>
 				</configuration>
 			</plugin>
+
 		</plugins>
 	</build>
-
 
 </project>
